Fixed build error by moving coordinate edit related classes from geoio library to guibase library.
