+++ libs/geodata/polygon/private/geodatapolygon_coordinateseditor.cpp
@@ -2,10 +2,10 @@
 #include "geodatapolygon_editcoordinatescommand.h"
 #include "../geodatapolygonabstractpolygon.h"
 
+#include <guibase/coordinates/coordinatesedit.h>
 #include <guibase/vtktool/vtkpointsutil.h>
 #include <guicore/base/iricmainwindowinterface.h>
 #include <guicore/pre/base/preprocessorwindowinterface.h>
-#include <geoio/coordinatesedit.h>
 #include <misc/iricundostack.h>
 
 #include <vtkPolygon.h>
