+++ libs/gridcreatingcondition/laplace/private/gridcreatingconditionlaplace_centerlinecoordinateseditor.cpp
@@ -2,11 +2,11 @@
 #include "gridcreatingconditionlaplace_impl.h"
 #include "../gridcreatingconditionlaplace.h"
 
+#include <guibase/coordinates/coordinatesedit.h>
 #include <guicore/base/iricmainwindowinterface.h>
 #include <guicore/pre/base/preprocessorwindowinterface.h>
 #include <guicore/project/projectdata.h>
 #include <guicore/project/projectmainfile.h>
-#include <geoio/coordinatesedit.h>
 #include <misc/iricundostack.h>
 
 #include <QString>
