+++ libs/guibase/coordinates/coordinatesedit.cpp
