+++ libs/guibase/coordinates/coordinateseditdialog.cpp
