+++ libs/guibase/coordinates/coordinateseditdialog.h
