+++ libs/guibase/coordinates/coordinateseditdialog.ui
