+++ libs/guibase/coordinates/private/coordinateseditdialogdelegate.h
